Sysinfo: Save file to %ProgramData% first, move to Desktop/Downloads later
Before, we just tried to place the Sysinfo file into `%USERPROFILE%\Desktop`.
This causes problems with OneDrive systems, where the Desktop folder is actually
in a different sub-directory ("OneDrive"). Additionally, the folder name is also
localized, meaning we can't just say `%USERPROFILE%\OneDrive\Desktop` all the time.

Thus, we now use SHGetKnownFolderPath to get the desktop path, and also fall back
to using the Downloads path (which can't be synced to OneDrive) if an error happens
there (for example when the OneDrive is full). And even if that fails and the whole
thing crashes, we still have the file with all the info in `%ProgramData%\24HS-Automator`

This also changes the file name from "DanielIsCool.txt" to "Sysinfo.txt", just to help users
in finding it.